# Add asynchronous support for plotting in tensorflow