--- a/rllab/algos/batch_polopt.py
+++ b/rllab/algos/batch_polopt.py
@@ -107,7 +107,6 @@ class BatchPolopt(RLAlgorithm):
     def start_worker(self):
         self.sampler.start_worker()
         if self.plot:
-            plotter.init_worker()
             plotter.init_plot(self.env, self.policy)
 
     def shutdown_worker(self):