Hi Ryan,
I have added the change for plotting using threads in tensorflow. I also corrected small issue with plotter used in theona.
Following algos were tested:
For theona:

1. DDPG : Using ddpg_cartpole.py example
2. NOP   : nop_carpole.py
3. CEM  : Using custom script with ddpg_cartpole.py as base
4. TRPO: Using trpo_cartpole_pickled.py , trpo_gym_cartpole.py and trpo_cartpole.py

For tensorflow:
1. VPG : Using vpg_cartpole.py example
2. TRPO: Using trpo_cartpole.py example

Performance analysis for tensorflow : 

Algorithm |   | Execution time (Without change) | Execution time (With change)
-- | -- | -- | --
VPG | REAL TIME | 20m42.579s | 8m51.536s
-- | USER LEVEL | 9m21.113s | 8m39.113s
-- | SYSTEM LEVEL | 0m13.016s | 0m13.254s
TRPO | REAL TIME | 16m4.019s | 4m4.585s
-- | USER LEVEL | 4m43.278s | 4m2.972s
-- | SYSTEM LEVEL | 0m14.059s | 0m13.738s