@ryanjulian 2018-03-14
Thanks for all your hard work! Just tested this on my machine and it seems to work great.

I think this PR is almost done. There are just a few pending issues.

1. Your newest edits are missing from the PR. I'm assuming this is an easy fix :).

1. The split between TensorFlow and Theano is not quite complete.
Most rllab launcher scripts use [run_experiment_lite](https://github.com/ryanjulian/rllab/blob/integration/scripts/run_experiment_lite.py#L79), as in [this example,](https://github.com/ryanjulian/rllab/blob/integration/examples/trpo_cartpole_pickled.py) to setup logging and plotting and run their algorithms in a structured way. You'll notice that `run_experiment_lite` calls `rllab.plotter.plotter.init_worker()` to setup the plotter worker. If I use `run_experiment_lite` from the TensorFlow tree, it will setup the wrong plotter.

1. Related to above, is that we need to set a convention for who calls plotter.init_worker() and stick to it. 
 I see three options
    1. **The current convention**. The user must always call `plotter.init_worker()` before running `algo.train(..., plot=True)`, either manually or using `run_experiment_lite`.  It makes it easy for the user to predict what processes and threads will be launched, because the user has to do it themselves. Remember that `algo.train()` can be called more than once in a given session. This is why `plotter.init_plot()` updates `env` and `policy` every time it is called.

    1. The algorithm calls `plotter.init_worker()` before attempting to plot. This makes it easy to support two different plotters, but adds redundant setup code to each algorithm and breaks the current assumptions in the Theano tree. **We would need to update the algorithms in the Theano tree to support this.** I think this is bad because it adds a setup step to each algorithm, and more than one algorithm can be trained in one session.

    1. `plotter` silently calls `init_worker()` the first time someone attempts to plot. This makes the plotter API easier to use, but the behavior of the program is now less predictable to the user. The user did not explicitly ask for more threads to be launched, and now `plot=True` has the side effect of launching a new thread/process. This is not so severe, since the rest of the software has similar side effects, but it must be considered.

If we do (i), we will need a different copy of `run_experiment_lite` for TensorFlow. If we do (iii) or (ii), we can keep the same `run_experiment_lite` for both. Lastly, if we do (ii), we will need to update some of algorithms to call `plotter.init_worker()`.

I say we go with (iii) and see if it works. This means we should remove the call to `plotter.init_worker()` from `run_experiment_lite`. This also means you should remove the call to `plotter.init_worker()` from `BatchPolopt`, since it should be called silently by init_plot.

```python
def init_worker(...):
    global queue
    if queue is None:    # this means the worker is not setup yet
        # normal setup steps

def init_plot(...):
    plotter.init_worker(...)  # note that there is no if statement
    #...
```