@raymoss 2018-03-14
I have added the commit with your suggestions. Once you give an approval, I will squash them into 1 single commit and we can submit it.